# optimize radius search by coherent memory access 