… python/test/core/test_nn.py