FixedRadiusSearch and HybridSearch use spatial hashing on GPU.
Sorting dataset points w.r.t hash id should make search faster since it enables coherent memory access on search time.
(Ref:  https://on-demand.gputechconf.com/gtc/2014/presentations/S4117-fast-fixed-radius-nearest-neighbor-gpu.pdf)

It turns out that the previous implementation of FixedRadiusSearch already has implemented to perform sorting in some way.
The only missing part was creating new dataset points re-arranged in the descending order of hash id.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/isl-org/open3d/3397)
<!-- Reviewable:end -->