@junha-l 2021-05-12
This is the benchmark result of with & without the optimization of this PR.

Without optimization (master branch) 
```
---------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations
---------------------------------------------------------------------------
BM_TestNNS_Radius/GPU_SORT/10          4.55 ms         4.55 ms          117
BM_TestNNS_Radius/GPU_SORT/50          10.9 ms         10.9 ms           65
BM_TestNNS_Radius/GPU_SORT/100         27.6 ms         27.6 ms           25
BM_TestNNS_Radius/GPU_SORT/200          109 ms          109 ms            6
BM_TestNNS_Radius/GPU_UNSORT/10        1.87 ms         1.87 ms          376
BM_TestNNS_Radius/GPU_UNSORT/50        2.88 ms         2.88 ms          244
BM_TestNNS_Radius/GPU_UNSORT/100       9.74 ms         9.74 ms           73
BM_TestNNS_Radius/GPU_UNSORT/200       55.7 ms         55.7 ms           13
```

With optimization (this branch)
```
---------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations
---------------------------------------------------------------------------
BM_TestNNS_Radius/GPU_SORT/10          1.97 ms         1.97 ms          318
BM_TestNNS_Radius/GPU_SORT/50          4.05 ms         3.96 ms          177
BM_TestNNS_Radius/GPU_SORT/100         15.7 ms         15.6 ms           45
BM_TestNNS_Radius/GPU_SORT/200          105 ms          105 ms            7
BM_TestNNS_Radius/GPU_UNSORT/10        1.91 ms         1.91 ms          371
BM_TestNNS_Radius/GPU_UNSORT/50        2.94 ms         2.94 ms          239
BM_TestNNS_Radius/GPU_UNSORT/100       10.0 ms         10.0 ms           71
BM_TestNNS_Radius/GPU_UNSORT/200       58.0 ms         57.9 ms           13
```

So it shows that sorting the dataset points w.r.t the hash bin id makes a positive impact on the performance **when we apply sorting** after the search (and also when the number of neighbors is relatively small). It doesn't lead to better performance when we do not sort (actually the running time gets slightly increased). 
I speculate that the current implementation introduces another **incoherent** memory access while mapping the neighbor indices to the order of original dataset points. (Because the re-ordered dataset points should not visible to the user).

I need more investigation into why it benefits when the sorting is applied.